+++ src/App.js
@@ -1,6 +1,6 @@
 import "./App.css";
 import {BrowserRouter, Route, Routes} from "react-router-dom";
-import PageLayout from "./PageLayout";
+import PageLayout from "./navigation/PageLayout";
 import Classes from "./classes/Classes";
 import Class from "./classes/class/Class";
