+++ src/navigation/NavBar.js
@@ -1,4 +1,4 @@
-import logo from "./logo.svg";
+import logo from "../logo.svg";
 import {Link} from "react-router-dom";
 
 function NavBar(props) {
