+++ src/navigation/PageLayout.js
@@ -1,6 +1,6 @@
 import NavBar from "./NavBar";
 import Footer from "./Footer";
-import NAV_LINKS from "./config/navbar.config";
+import NAV_LINKS from "../config/navbar.config";
 import {Outlet} from "react-router-dom";
 
 function PageLayout() {
